Update to bsnes v089r07 release.
Changelog:
* EpsonRTC emulation improved further (stop/pause blocks IRQs,
verified secondhi >= 3 triggers 30-second adjust (even on invalid
BCD), second-changed flag is mirrored to
minute+hour+day+month+weekday, improved busy timing, etc.)
* SharpRTC rewritten, works like EpsonRTC now in that it has its own
timing thread and ticks with the emulation
* won't attempt to read from an unopen file stream now (I think this
is what was crashing Sufami Turbo without SRAM?)
* added Tools -> Synchronize Time option below load/save state
options. Only appears when you play a game with an emulated RTC chip

Just realized that I used 125ms for the 30-second adjust instead of
125us, so I'll fix that in the next WIP.
Aside from that, this is as good as the emulation is going to get.
There's still a couple of absolutely psychopathic edge cases that are
just too damn difficult to simulate.
So that leaves us with data port control + decompression status
registers to investigate before SPC7110 will be finished.

I also still need to redo the BS-X RTC ... I'm not even sure if there
is a time-set for it, right now I just return the current system time
because of that
And then lastly, there's the Game Boy RTC used by one of the memory
mappers. Not sure if there are NES/GBA RTCs or not.

[No archive available]