Update to bsnes v089r06a release.
I had some bugs in r07 that I couldn't track down, DKJM2's clock was
getting all out of sync.
So I just reverted to r05, blew away both RTCs entirely, and wrote
them cleanly from scratch (obviously looking off the old code.) A bit
extreme, but it worked.
I believe I found the error in the process, day and month were
resetting the counter to 0 instead of 1, it wasn't handling leap year,
etc.
While I was at it, I fixed the day-of-week calculation. The SharpRTC
epoch is actually 1000-01-01, and not 1900-01-01.
I'm sure you guys will be really happy that if it ever becomes 1000AD
again and you're playing a ROM hack that uses the SharpRTC and relies
on its weekday value that it will show the correct day now ...
Kind of a pain to compute, but nothing compared to the seventh circle
of hell that was my IBM dBase III Julian<>Gregorian conversion
functions :/
Also found a few bugs in the Epson code this way. And I moved the
round seconds actions and flag clear to +125us after flag set.

So, if you had the old r06 or r07, please delete those.

Unfortunately, this took all of my energy today, so the file names
inside manifest changes will have to be in the next WIP.

EDIT: ran a diff against old r07 and new r06.
* added if(days == 31) case twice in EpsonRTC::tick_day()
* forgot weekday = 0; in SharpRTC::load()
* need to move the cartridge+cheat objects up in sfc/Makefile again
* System::init() needs assert(interface != nullptr /* not 0 */)