Update to bsnes v089r07a release.
Not even purify makes compatible images for this WIP.
Unless you want to figure it out yourself, I'd suggest waiting for an
updated tool before using subsequent WIPs.

Changelog:
* MSU1 initializes data port + audio track to 0
* MSU1 implements audio track error flag on $2000.d3
* manifest.xml now controls file names for cartridge folders ...
mostly

Regressions:
* Super Game Boy support is broken
* Sufami Turbo support is broken

So, basically Emulator::Interface() now has:
void load(const string &manifest);
void save();

The first one will analyze the manifest, and call all the ROM + RAM
loadRequest() commands necessary to run the game.
The second one will call saveRequest() commands on all writable and
non-volatile storage (basically if it's a RAM type and has a filename
specified, it gets saved to disk.)
save() shrinks the size of Emulator::Interface() by hiding information
one is unlikely to care about. It also makes it much easier to save.
The core auto-calls this when you unload a game as well. So the only
time you ever have to worry about it is if you want to save RAM files
mid-game (in case you want to do periodic backups in case of a crash.)

BSX only seems to work by some odd fluke, but all of the sub-carts
have to be rethought. So that's going to be a pain in the ass for
sure.
We also may as well address MSU1 file naming as well.

This change also makes a lot of manifest.xml markup more interesting.
Do we claim BWRAM is the same thing as cartridge RAM? What about SA-1
ROM? Is that the same as cartridge ROM?
Or how about the SPC7110? Do we have a <cartridge><rom> tag, or do we
put the ROM names inside <cartridge><spc7110><mcu><rom>+<ram>?

I like the idea of having separate files for each tag. But right now,
some things end up mapped both inside and outside of special chips, so
a lot of work needs to be done to harmonize everything.