Update to bsnes v089r08 release.
Changelog:
* Super Game Boy, BS-X Satellaview and Sufami Turbo cartridges all
load manifests that specify their file names, and they all work
* Sufami Turbo can now properly handle carts without RAM, or empty
slots entirely
* Emulator::Interface structures no longer specify any file names,
ever
* exposed "capability.(cheats,states)" now. So far, this just means
the GBA doesn't show the cheat editor, since it doesn't support cheat
codes yet
* as such, state manager and cheat editor windows auto-hide (may be a
tiny bit inconvenient, but it makes not having to sync them or deal
with input when no cart is loaded easier)
* added "AbsoluteInput" type, which returns mouse coordinates from
-32767,-32767 (top left) to +32767,+32767 (bottom right) or
-32768,-32768 (offscreen)

AbsoluteInput is just something I'm toying with. Idea is to support eg
Super Scope or Justifier, or possibly some future Famicom controllers
that are absolute-indexed. The coordinates are scaled, so the bigger
your window, the more precise they are. But obviously you can't get
more precise than the emulated system, so 1x scale will behave the
same anyway. I haven't hooked it up yet, need to mess with the idea of
custom cursors via phoenix for that first. Also not sure if it will
feel smoother or not ... if you resize the window, your mouse will
seem to move slower. Still, not having to capture the mouse for SS/JS
may be nicer yet. But we'll see ... just experimenting for now.