Update to bsnes v089r17 release.
This implements the spec from the XML part 3 thread
(viewtopic.php?f=16&t=2998)
It's also propagated the changes to nall and purify, so you can test
this one.

This is basically it, after years of effort I feel I finally have a
fully consistent and logical XML board format.
The only things left to change will be: modifications if emulation
turns out to be incorrect (eg we missed some MMIO mirrors, or mirrored
too much), and new additions.

And of course, I'm giving it a bit of time for good arguments against
the format.

Other than that, this release removes linear_vector and
pointer_vector, as vector is better than linear_vector and I've never
used pointer_vector.
vector also gets move(), which is a way to use move-semantics across
types. It lets you steal the underlying memory pool, effectively
destroying the vector without a copy.
This works really nicely with the move for read() functions to return
vector<uint8> instead of taking (uint8_t*&, unsigned&) parameters.