Update to higan v091r05 release.
http://byuu.org/files/wip/higan_v091r05.tar.xz

Super Famicom mapping system has been reworked as discussed with the
mask= changes. offset becomes base, mode is gone.
Also added support for comma-separated fields in the address fields,
to reduce the number of map lines needed.

    <?xml version="1.0" encoding="UTF-8"?>
    <cartridge region="NTSC">
      <superfx revision="2">
        <rom name="program.rom" size="0x200000"/>
        <ram name="save.rwm" size="0x8000"/>
        <map id="io" address="00-3f,80-bf:3000-32ff"/>
        <map id="rom" address="00-3f:8000-ffff" mask="0x8000"/>
        <map id="rom" address="40-5f:0000-ffff"/>
        <map id="ram" address="00-3f,80-bf:6000-7fff" size="0x2000"/>
        <map id="ram" address="70-71:0000-ffff"/>
      </superfx>
    </cartridge>


Or in BML:
    cartridge region=NTSC
      superfx revision=2
        rom name=program.rom size=0x200000
        ram name=save.rwm size=0x8000
        map id=io address=00-3f,80-bf:3000-32ff
        map id=rom address=00-3f:8000-ffff mask=0x8000
        map id=rom address=40-5f:0000-ffff
        map id=ram address=00-3f,80-bf:6000-7fff size=0x2000
        map id=ram address=70-71:0000-ffff


As a result of the changes, old mappings will no longer work. The
above XML example will run Super Mario World 2: Yoshi's Island.
Otherwise, you'll have to write your own.

All that's left now is to work some sort of database mapping system
in, so I can start dumping carts en masse.

The NES changes that FitzRoy asked for are mostly in as well.

Also, part of the reason I haven't released a WIP ... but fuck it, I'm
not going to wait forever to post a new WIP.

I've added a skeleton driver to emulate Campus Challenge '92 and
Powerfest '94. There's no actual emulation, except for the stuff I can
glean from looking at the pictures of the board. It has a DSP-1 (so
SR/DR registers), four ROMs that map in and out, RAM, etc.

I've also added preliminary mapping to upload high scores to a
website, but obviously I need the ROMs first.

Still trying to get Rick or JJ to lend me the ROMs. I'm also hoping to
emulate NWC '90 and/or CC '91 (if I can get it dumped) to show that I
can do the SNES variants. So let's keep this between us here, please.