Update to higan v091r06 release.
This release adds initial database support.

The way it works is you can now load game folders as you always have,
or you can load a game file.
If you load a game file, it tries to create a game folder for you by
looking up the file's sha256 in a database.
If it can't find it, sorry, the game won't play. I'm not hooking up
the oldschool "make up a manifest" code here.
The easiest way to handle this is to get me every game so I can dump
it and add it to the database :D

The database entries are complete entries that can be copied directly.
So it describes the board, the information, file layout, etc.
That'll be what comes with higan releases in the future.

Internally, I'm separating the information and board descriptions, and
will use a tool to merge the two together.

Here's a current database copy, with one game in it. Still hammering
out some details, but it's mostly how it's going to look.

    cartridge region=NTSC
      board type=1CB5B-20
        superfx revision=2
          rom name=program.rom size=0x200000
          ram name=save.rwm size=0x8000
          map id=io address=00-3f,80-bf:3000-32ff
          map id=rom address=00-3f:8000-ffff mask=0x8000
          map id=rom address=40-5f:0000-ffff
          map id=ram address=00-3f,80-bf:6000-7fff size=0x2000
          map id=ram address=70-71:0000-ffff
      information
        name:   Super Mario World 2 - Yoshi's Island (SNS) (1.1)
        title:  Super Mario World 2: Yoshi's Island
        sha256:
    bd763c1a56365c244be92e6cffefd318780a2a19eda7d5baf1c6d5bd6c1b3e06
        board:  SHVC-1CB5B-20
        rom:    0x200000
        ram:    0x8000
      layout
        file name=program.rom size=0x200000