Update to higan v091r09 release.
I'd appreciate it if you guys wouldn't mind testing out the database
functionality.

Save this file as database.bml (remove the date) inside
~/.config/higan/Super Famicom.sfc/ or %APPDATA%/higan/Super
Famicom.sfc/
http://byuu.org/snes/database/database_2012-10-21.bml

Now load any of the 20 games in the database from the file dialog.
They need to be named *.sfc, have no copier header, and have firmware
appended (for Mario Kart only so far.)

If anyone actually does test it, please let me know how it goes for
you and what you think.
Note that future versions of higan will have the database.bml file
included with the release.
I'm undecided if I want to support ZIP/SMC/copier headers. It's not as
bad since it's being stripped off first anyway.

You'll note that I'm outputting a 32-byte file named "sha256", which
is a textual SHA256 hash to find the game in the database.
Also outputting the manifest.bml file for now, which is what higan is
using.
Plan is that if we can lock down on these file names (and I really
think we can), we can omit the manifest, and just pull it from the
database upon load.
Then manifests will only be needed for unofficial and unverified
software.