Update to higan v092 release.
The first official release of higan has been posted. higan is the new name for bsnes, and it continues with the latter's version numbering.
Note that as of now, bsnes still exists. It's a module distributed inside of higan. bsnes is now specific to my SNES emulator.
Due to last minute changes to the emulator interface, and missing support in ananke, I wasn't able to include Cydrak's Nintendo DS emulator dasShiny in this build, but I hope to do so in the next release.
higan is very different from most emulators. I strongly recommend that you read the higan user guide, even if you are already familiar with bsnes and/or higan.
Also, for users of bsnes v091, please note that the game folder format has changed. Unfortunately, I do not yet have a tool ready to update your game folders. So you have two options: delete all games from ~/Emulation, or wait for a new version of purify before upgrading from v091 to v092. For users of bsnes v090 and earlier, you should be able to jump right in after having read the user guide.