Update to higan v092r05 release.
This release should be polished enough for a general release.

Anyone with a real, clean Mac up for posting compiled binaries?
Preferably compile with "make profile=balanced"
In fact, I'd like it if someone were willing to host a "higan for Mac"
page, with binaries of each of the latest releases.
Only really needed for major official releases, but it'd be preferable
to have the builds updated as soon as possible after I post new
builds.

Changelog:
* no more keyboard chimes when pressing keys
* status bar added, fully functional
* Label::minimumSize() takes frame into account (but note a few places
hard-code raw Font::size(), so a few text labels are still clipped)
* resizing the main window looks smooth regardless of whether a game
is running or not
*** currently, resizing the window pauses the emulation. Allowing it
to run the main loop was lagging out the window resize process too
much to be worth it

Additional OS X integration enhancements:
* closing the main window unloads the current game, but does not quit
the application (quit via the main menu or the dock menu)
* clicking the icon in the dock will (re)display the main menu