Update to higan v092r08 release.
Changelog:
* fixed cartridge load window focus on Windows
* lots of updates to nall, ruby and phoenix
* ethos and Emulator::Interface updated from "foo &bar" to "foo& bar"
syntax (work-in-progress)

Before I had mixed the two ways to declare variables/arguments all
over the place, so the goal is to unify them all for consistency. So
the changelog for this release will be massive (750KB >.>) due to the
syntax change. Yeah, that's what I spent the last three days working
on ...

I don't suppose anyone would mind helping me look over the diff for
any inadvertent typos? :/
Would be a real bitch to track down if I managed to mistype a single
character somewhere.