Update to higan v093r02 release.
Changelog:
* nall: fixed major memory leak in string class
* ruby: video shaders support #define-based settings now
* phoenix/GTK+: support > 256x256 icons for window / task bar / alt-
tab
* sfc: remove random/ and config/, merge into system/
* ethos: delete higan.png (48x48), replace with higan512.png (512x512)
as new higan.png
* ethos: default gamma to 100% (no color adjustment)
* ethos: use "Video Shaders/Display Emulation/" instead of "Video
Shaders/Emulation/"
* use g++ instead of g++-4.7 (g++ -v must be >= 4.7)
* use -std=c++11 instead of -std=gnu++11
* applied a few patches from Debian upstream to make their packaging
job easier

So because colors are normalized in GLSL, I won't be able to offer
video shaders absolute color literals. We will have to perform basic
color conversion inside the core.

As such, the current plan is to create some sort of Emulator::Settings
interface. With that, I'll connect an option for color correction,
which will be on by default. For FC/SFC, that will mean gamma
correction (darker / stronger colors), and for GB/GBC/GBA, it will
mean simulating the weird brightness levels of the displays. I am
undecided on whether to use pea soup green for the GB or not. By not
doing so, it'll be easier for the display emulation shader to do it.