Update to higan v093r11 release.
Changelog:
* GBA: SOUND_CTL_H is readable, fixes sound effects in Mario&Luigi
Superstar Saga [Cydrak] (note: game is still unplayable due to other
bugs)
* phoenix/Windows: workaround for Win32 API ListView bug, fixes slot
loading behavior
* ruby: added udev driver for Linux with rumble support, and added
rumble support to existing RawInput driver for XInput and DirectInput
* ethos: added new "Rumble" mapping to GBA input assignment, use it to
tell higan which controller to rumble (clear it to disable rumble)
* GBA: Game Boy Player rumble is now fully emulated
* core: added new normalized raw-color palette mode for Display
Emulation shaders

The way rumble was added to ethos was somewhat hackish. The support
doesn't really exist in nall.

I need to redesign the entire input system, but that's not a change I
want to make so close to a release.