Update to higan v094r07 release.
Changelog for loki:
* added command aliases (match with * [sorry, regex lib isn't
available everywhere yet], replace with {1}+)
* added command hotkeys
* added window geometry saving
* added save state support
* added power/reset commands
* added an input manager, so you can remap keys (limiting it to the
keyboard for now though)

The combination of aliases and hotkeys really makes things shine. Save
states will temporarily disable your breakpoints (run/step are
technically temporary breakpoints) so as to ensure the state is
captured at a good time. In practice, this should pose about as much
of a problem as higan desyncing and breaking when capturing states ...
should be exceedingly rare to ever even notice this behavior at all,
with 99.9% of state captures happening in half an instruction
boundary. But still, keep it in mind, as you might see the CPU step
one instruction ahead. Tracing and usage map functionality is still
enabled during state synchronization.

So at this point, I have 100% of the essential stuff in. All that's
left now is to add polish / wishlist features like bass and mosaic
integration.